@@ -21,9 +21,9 @@
     "less": "^3.11.3",
     "less-loader": "^6.1.0",
     "lodash": "^4.17.15",
-    "react": "^16.13.1",
+    "react": "^17.0.1",
     "react-app-rewired": "^2.1.6",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.1",
     "react-infinite-scroller": "^1.2.4",
     "react-redux": "^7.2.0",
     "react-router": "^5.2.0",
@@ -44,13 +44,14 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "10.3.0",
     "@testing-library/user-event": "^7.1.2",
+    "@types/extract-files": "^8.1.0",
     "@types/graphql": "^14.5.0",
     "@types/jest": "^24.0.0",
     "@types/jest-when": "^2.7.1",
     "@types/lodash": "^4.14.155",
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@types/react-dom": "^17.0.0",
     "@types/react-infinite-scroller": "^1.2.1",
     "@types/react-redux": "^7.1.9",
     "@types/react-router": "^5.1.7",
