@@ -1,15 +1,12 @@
-import { TYPES } from '@stringsync/di';
-import { injectable, inject } from 'inversify';
-import { TagRepo } from '@stringsync/repos';
+import { inject, injectable } from '@stringsync/di';
 import { Tag } from '@stringsync/domain';
+import { REPOS_TYPES, TagRepo } from '@stringsync/repos';
+
+const TYPES = { ...REPOS_TYPES };
 
 @injectable()
 export class TagService {
-  tagRepo: TagRepo;
-
-  constructor(@inject(TYPES.TagRepo) tagRepo: TagRepo) {
-    this.tagRepo = tagRepo;
-  }
+  constructor(@inject(TYPES.TagRepo) public tagRepo: TagRepo) {}
 
   async find(id: string): Promise<Tag | null> {
     return await this.tagRepo.find(id);
