@@ -1,15 +1,12 @@
-import { injectable, inject } from 'inversify';
-import { TaggingRepo } from '@stringsync/repos';
-import { TYPES } from '@stringsync/di';
+import { inject, injectable } from '@stringsync/di';
 import { Tagging } from '@stringsync/domain';
+import { REPOS_TYPES, TaggingRepo } from '@stringsync/repos';
+
+const TYPES = { ...REPOS_TYPES };
 
 @injectable()
 export class TaggingService {
-  taggingRepo: TaggingRepo;
-
-  constructor(@inject(TYPES.TaggingRepo) taggingRepo: TaggingRepo) {
-    this.taggingRepo = taggingRepo;
-  }
+  constructor(@inject(TYPES.TaggingRepo) public taggingRepo: TaggingRepo) {}
 
   async bulkCreate(taggings: Array<Omit<Tagging, 'id'>>): Promise<Tagging[]> {
     return await this.taggingRepo.bulkCreate(taggings);
