@@ -1,7 +1,7 @@
 import { Command, flags } from '@oclif/command';
-import { PROJECT_ARG, ROOT_PATH } from '../util/constants';
-import { getDockerComposeFile } from '../util';
 import { spawn } from 'child_process';
+import { getDockerComposeFile } from '../util';
+import { DOCKER_PATH, PROJECT_ARG } from '../util/constants';
 
 export default class Down extends Command {
   static description = 'Turns down a docker-compose environment.';
@@ -16,7 +16,7 @@ export default class Down extends Command {
     const { args } = this.parse(Down);
 
     spawn('docker-compose', ['-p', args.project, '-f', getDockerComposeFile(args.project), 'down', '--volumes'], {
-      cwd: ROOT_PATH,
+      cwd: DOCKER_PATH,
       stdio: 'inherit',
     });
   }
