@@ -1,7 +1,7 @@
 import { Command, flags } from '@oclif/command';
 import { execSync } from 'child_process';
 import { getDockerComposeFile } from '../util';
-import { ROOT_PATH } from '../util/constants';
+import { DOCKER_PATH, ROOT_PATH } from '../util/constants';
 
 export default class Test extends Command {
   static description = 'Run all of the StringSync tests.';
@@ -15,8 +15,7 @@ export default class Test extends Command {
     {
       name: 'project',
       required: true,
-      default: 'server',
-      options: ['server', 'web'],
+      options: ['api', 'web'],
     },
     { name: 'cmd', required: false },
   ];
@@ -24,7 +23,7 @@ export default class Test extends Command {
   async run() {
     const { flags, args, argv } = this.parse(Test);
 
-    execSync(['./bin/ss', 'build'].join(' '), {
+    execSync(['./bin/ss', 'build', '-d'].join(' '), {
       cwd: ROOT_PATH,
       stdio: 'inherit',
     });
@@ -51,7 +50,7 @@ export default class Test extends Command {
           .filter((part) => part)
           .join(' '),
         {
-          cwd: ROOT_PATH,
+          cwd: DOCKER_PATH,
           stdio: 'inherit',
         }
       );
