@@ -1,6 +1,6 @@
 import { Command, flags } from '@oclif/command';
-import { ROOT_PATH, PROJECT_ARG, getDockerComposeFile } from '../util';
 import { execSync, spawn } from 'child_process';
+import { DOCKER_PATH, getDockerComposeFile, PROJECT_ARG } from '../util';
 
 export default class Up extends Command {
   static description = 'Spins up a development environment';
@@ -15,12 +15,12 @@ export default class Up extends Command {
     const { args } = this.parse(Up);
 
     if (args.project === 'web') {
-      execSync('yarn web dev', { cwd: ROOT_PATH, stdio: 'inherit' });
+      execSync('yarn web dev', { cwd: DOCKER_PATH, stdio: 'inherit' });
       this.exit();
     }
 
     spawn('docker-compose', ['-p', args.project, '-f', getDockerComposeFile(args.project), 'up', '--build', '-d'], {
-      cwd: ROOT_PATH,
+      cwd: DOCKER_PATH,
       stdio: 'inherit',
     });
   }
