@@ -4,7 +4,9 @@ import { Project } from './types';
 // dirname is bin, since the commands will be run from there
 export const ROOT_PATH = path.join(__dirname, '..', '..', '..', '..');
 
-export const PROJECTS: Project[] = ['main', 'e2e', 'server', 'web'];
+export const DOCKER_PATH = path.join(ROOT_PATH, 'docker');
+
+export const PROJECTS: Project[] = ['main', 'e2e', 'api', 'web'];
 
 export const PROJECT_ARG = {
   name: 'project',
