@@ -8,9 +8,9 @@ module.exports = {
     '^.+\\.tsx?$': 'ts-jest',
   },
   resetModules: false,
-  testPathIgnorePatterns: ['node_modules', 'web'],
+  testPathIgnorePatterns: ['node_modules', 'web', 'build', 'dist'],
   testEnvironment: '<rootDir>/ServerTestEnvironment.js',
   roots: ['<rootDir>/packages'],
-  setupFilesAfterEnv: ['jest-extended', '<rootDir>/packages/graphql/src/testing/jest.setup.ts'],
+  setupFilesAfterEnv: ['jest-extended', '<rootDir>/packages/api/src/testing/jest.setup.ts'],
   watchPlugins: ['jest-watch-typeahead/filename', 'jest-watch-typeahead/testname'],
 };
