@@ -1,5 +1,7 @@
-export * from './types';
+export * from './Base64';
+export * from './compose';
+export * from './ctor';
 export * from './identity';
 export * from './noop';
-export * from './compose';
-export * from './Base64';
+export * from './types';
+export * from './wait';
