@@ -1,4 +1,3 @@
 export * from './configFactory';
-export * from './getContainerConfig';
-export * from './getWebConfig';
+export * from './constants';
 export * from './types';
