@@ -1,4 +1,5 @@
-export * from './user';
-export * from './tag';
 export * from './notation';
+export * from './tag';
 export * from './tagging';
+export * from './testing';
+export * from './user';
