@@ -1,61 +1,61 @@
 import { UnknownError } from '@stringsync/common';
+import { injectable } from '@stringsync/di';
 import { JobsOptions, Queue, QueueScheduler, Worker } from 'bullmq';
-import { injectable } from 'inversify';
 import * as uuid from 'uuid';
 
 @injectable()
 export abstract class Job<T = undefined> {
-  private _queue: Queue<T> | null = null;
-  private _worker: Worker<T> | null = null;
-  private _scheduler: QueueScheduler | null = null;
+  private queue: Queue<T> | null = null;
+  private worker: Worker<T> | null = null;
+  private scheduler: QueueScheduler | null = null;
 
   protected abstract createQueue(): Queue<T>;
   protected abstract createWorker(): Worker<T>;
   protected abstract createScheduler(): QueueScheduler | null;
 
   setupWorker() {
-    if (!this._scheduler) {
-      this._scheduler = this.createScheduler();
+    if (!this.scheduler) {
+      this.scheduler = this.createScheduler();
     }
-    if (!this._worker) {
-      this._worker = this.createWorker();
+    if (!this.worker) {
+      this.worker = this.createWorker();
     }
   }
 
   setupQueue() {
-    if (!this._queue) {
-      this._queue = this.createQueue();
+    if (!this.queue) {
+      this.queue = this.createQueue();
     }
   }
 
   async teardownWorker() {
     const promises = [];
 
-    if (this._scheduler) {
-      promises.push(this._scheduler.close());
+    if (this.scheduler) {
+      promises.push(this.scheduler.close());
     }
-    if (this._worker) {
-      promises.push(this._worker.close());
+    if (this.worker) {
+      promises.push(this.worker.close());
     }
 
     await Promise.all(promises);
 
-    this._scheduler = null;
-    this._worker = null;
+    this.scheduler = null;
+    this.worker = null;
   }
 
   async teardownQueue() {
-    if (this._queue) {
-      await this._queue.close();
+    if (this.queue) {
+      await this.queue.close();
     }
   }
 
   async enqueue(data: T, opts?: JobsOptions) {
-    if (!this._queue) {
+    if (!this.queue) {
       throw new UnknownError(`must setup queue first`);
     }
     const jobId = await this.getJobId();
-    this._queue.add(jobId, data, opts);
+    this.queue.add(jobId, data, opts);
   }
 
   protected async getJobId() {
@@ -67,10 +67,10 @@ export abstract class Job<T = undefined> {
   }
 
   private async exists(jobId: string) {
-    if (!this._queue) {
+    if (!this.queue) {
       throw new UnknownError(`must setup queue first`);
     }
-    const job = await this._queue.getJob(jobId);
+    const job = await this.queue.getJob(jobId);
     return typeof job !== 'undefined';
   }
 }
