@@ -1,4 +1,7 @@
-export * from './types';
-export * from './sequelize';
-export * from './util';
 export * from './factory';
+export * from './REPOS';
+export * from './REPOS_CONFIG';
+export * from './REPOS_TYPES';
+export * from './sequelize';
+export * from './types';
+export * from './util';
