@@ -1,11 +1,10 @@
+import { NotationModel, TaggingModel } from '@stringsync/db';
+import { injectable } from '@stringsync/di';
+import { Notation, Tagging } from '@stringsync/domain';
+import Dataloader from 'dataloader';
 import { groupBy, mapValues } from 'lodash';
 import { NotationLoader } from '../../types';
-import Dataloader from 'dataloader';
-import { TYPES } from '@stringsync/di';
-import { NotationModel, UserModel, TaggingModel, TagModel } from '@stringsync/db';
-import { inject, injectable } from 'inversify';
-import { Notation, Tagging } from '@stringsync/domain';
-import { alignOneToMany, alignOneToOne, ensureNoErrors, alignManyToMany } from '../../util';
+import { alignManyToMany, alignOneToMany, alignOneToOne, ensureNoErrors } from '../../util';
 
 @injectable()
 export class NotationSequelizeLoader implements NotationLoader {
