@@ -1,19 +1,14 @@
+import { injectable } from '@stringsync/di';
 import { S3 } from 'aws-sdk';
 import { Stream } from 'stream';
-import { BlobStorage, S3Config } from './types';
+import { BlobStorage } from './types';
 
+@injectable()
 export class S3Storage implements BlobStorage {
-  static create(config: S3Config): S3Storage {
-    const s3 = new S3();
-    return new S3Storage(s3, config.domainName);
-  }
-
   s3: S3;
-  domainName: string;
 
-  constructor(s3: S3, domainName: string) {
-    this.s3 = s3;
-    this.domainName = domainName;
+  constructor() {
+    this.s3 = new S3();
   }
 
   async put(filepath: string, bucket: string, readStream: Stream) {
