@@ -3,4 +3,6 @@ export * from './cache';
 export * from './logger';
 export * from './mailer';
 export * from './message-queue';
-export * from './redis';
+export * from './UTIL';
+export * from './UTIL_CONFIG';
+export * from './UTIL_TYPES';
