@@ -6,10 +6,10 @@ export type Message = {
   id: string;
   body: string;
   recieptHandle: string;
-  queueName: string;
+  queueUrl: string;
 };
 
 export interface MessageQueue {
-  get(queueName: string): Promise<Message | null>;
+  get(queueUrl: string): Promise<Message | null>;
   ack(message: Message): Promise<void>;
 }
